+++ spyro/utils/error_management.py
@@ -120,7 +120,7 @@ def clean_inst_num(data_arr):
     data_arr : `array`
         An array with null or positive components.
     """
-    type_data_structure_error("data_arr", data_arr, "ndarray", ("float", "int"))
+    type_data_structure_error("data_arr", data_arr, "array", ("float", "int"))
     data_arr[where(isnan(data_arr) | isinf(data_arr) | (data_arr < 0.0))] = 0.0
     return data_arr
 
@@ -314,12 +314,12 @@ def type_data_structure_error(par_name, par_value, expected_type,
     """
 
     value_parameter_error("expected_type", expected_type,
-                          ["dict", "list", "tuple", "ndarray"])
+                          ["dict", "list", "tuple", "array"])
 
     parameter_map = {"dict": dict,
                      "list": list,
                      "tuple": tuple,
-                     "ndarray": ndarray}
+                     "array": ndarray}
 
     element_map = {"int": int,
                    "float": float,
